+++ benchmark/forest.jl
@@ -27,12 +27,12 @@ end
 
 function forest_model_step_array!(forest)
     for pos in positions(forest, :random)
-        np = agents_in_pos(pos, forest)
+        ids = ids_in_position(pos, forest)
         ## the position is empty, maybe a tree grows here
-        if length(np) == 0
+        if length(ids) == 0
             rand() ≤ forest.p && add_agent!(pos, forest, true)
         else
-            tree = forest[np[1]] # by definition only 1 agent per position
+            tree = forest[ids[1]] # by definition only 1 agent per position
             if tree.status == false  # if it is has been burning, remove it.
                 kill_agent!(tree, forest)
             else