+++ src/Agents.jl
@@ -28,7 +28,8 @@ include("simulations/sample.jl")
 # 4.0 Depreciations
 @deprecate space_neighbors nearby_ids
 @deprecate node_neighbors nearby_positions
-@deprecate get_node_contents agents_in_pos
+@deprecate get_node_contents ids_in_position
+@deprecate get_node_agents agents_in_position
 @deprecate pick_empty random_empty
 @deprecate find_empty_nodes empty_positions
 @deprecate has_empty_nodes has_empty_positions